reorganize implementations
