+++ app/agents/__init__.py
@@ -1,0 +1,4 @@
+from app.agents.cohere_agent import CohereAgent
+from app.agents.fake_agent import FakeAgent
+
+__all__ = ["FakeAgent", "CohereAgent"]
